split Tests into UIGestureRecognizer and UIPanGestureRecognizer
