+++ UIKitTests/Gestures/UIGestureRecognizerTests.swift
@@ -2,7 +2,7 @@
 //  UIGestureRecognizerTests.swift
 //  UIKitTests
 //
-//  Created by flowing erik on 22.09.17.
+//  Created by flowing erik on 23.10.17.
 //  Copyright © 2017 flowkey. All rights reserved.
 //
 
@@ -13,25 +13,25 @@ fileprivate class TestPanGestureRecognizer: UIPanGestureRecognizer {
     let stateCancelledExpectation: XCTestExpectation?
     let stateEndedExpectation: XCTestExpectation?
 
-    override public var state: UIGestureRecognizerState {
-        didSet {
-            switch state {
-            case .ended: stateEndedExpectation?.fulfill()
-            case .cancelled: stateCancelledExpectation?.fulfill()
+    fileprivate init (cancelledExp: XCTestExpectation? = nil, endedExp: XCTestExpectation? = nil) {
+        stateCancelledExpectation = cancelledExp
+        stateEndedExpectation = endedExp
+        super.init()
+        self.onStateChanged = {
+            switch self.state {
+            case .ended: self.stateEndedExpectation?.fulfill()
+            case .cancelled: self.stateCancelledExpectation?.fulfill()
             default: break
             }
         }
     }
-
-    fileprivate init (cancelledExp: XCTestExpectation? = nil, endedExp: XCTestExpectation? = nil){
-        stateCancelledExpectation = cancelledExp
-        stateEndedExpectation = endedExp
-    }
 }
 
-class UIGestureRegognizerTests: XCTestCase {
+class UIGestureRecognizerTests: XCTestCase {
+
     var mockTouch: UITouch!
     let mockView = UIView(frame: CGRect(origin: .zero, size: CGSize(width: 100, height: 100)))
+
     override func setUp() {
         mockTouch = UITouch(
             at: CGPoint(x: 0, y: 0),
@@ -39,7 +39,12 @@ class UIGestureRegognizerTests: XCTestCase {
             touchId: 0
         )
     }
-
+    
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        super.tearDown()
+    }
+    
     func testPanGestureRecognizerStateEnded() {
         let endedExpectation = expectation(description: "State was ended")
         let pgr = TestPanGestureRecognizer(endedExp: endedExpectation)
@@ -52,7 +57,11 @@ class UIGestureRegognizerTests: XCTestCase {
         XCTAssert(pgr.state == .changed)
 
         pgr.touchesEnded([mockTouch], with: UIEvent())
-        wait(for: [endedExpectation], timeout: 1)
+
+        // expect that state was .ended at some point
+        wait(for: [endedExpectation], timeout: 2)
+
+        // state should transition to .possible after being .ended
         XCTAssert(pgr.state == .possible)
     }
 
@@ -68,43 +77,11 @@ class UIGestureRegognizerTests: XCTestCase {
         XCTAssert(pgr.state == .changed)
 
         pgr.touchesCancelled([mockTouch], with: UIEvent())
+
+        // expect that state was .cancelled at some point
         wait(for: [cancelledExpectation], timeout: 1)
+
+        // state should transition to .possible after being .cancelled
         XCTAssert(pgr.state == .possible)
-    }
-
-    func testPanGestureVelocity () {
-        let touchPositionDiff: CGFloat = 50
-        let timeInterval = 1.0
-
-        let pgr = UIPanGestureRecognizer()
-        let velocityExp = expectation(description: "velocity is as expected")
-
-        self.mockTouch.previousPositionInView = CGPoint(x: 0, y: 0)
-        self.mockTouch.positionInView = CGPoint(x: touchPositionDiff, y: 0)
-        pgr.touchesBegan([mockTouch], with: UIEvent())
-        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + timeInterval) {
-            pgr.touchesMoved([self.mockTouch], with: UIEvent())
-            let velocityX = pgr.velocity(in: self.mockView).x
-            let expectedVelocityX: CGFloat = touchPositionDiff / CGFloat(timeInterval)
-
-            print(velocityX, expectedVelocityX)
-            if velocityX.isRoundAbout(to: expectedVelocityX, percentalAccuracy: 5.0) {
-                velocityExp.fulfill()
-            }
-        }
-
-        wait(for: [velocityExp], timeout: 1.1)
-    }
-}
-
-fileprivate extension CGFloat {
-    func isRoundAbout(to value: CGFloat, percentalAccuracy: Double) -> Bool {
-        let min = Double(value) - ((Double(value) * percentalAccuracy) / 100)
-        let max = Double(value) + ((Double(value) * percentalAccuracy) / 100)
-        let result = (min ..< max)~=(Double(self)) // if in range
-        if (result == false) {
-            fatalError(String(describing: self) + "is not in range between " + String(describing: min) + " and " + String(describing: max))
-        }
-        return result
     }
 }
